DEX-539 Merge 1.1.x into master
Compilation fixed in dex-it/test (WIP)
