+++ dex-it/src/test/scala/com/wavesplatform/it/sync/PostgresHistoryDatabaseTestSuite.scala
@@ -1,184 +1,152 @@
 package com.wavesplatform.it.sync
 
-import java.net.InetAddress
 import java.sql.{Connection, DriverManager}
 
-import akka.http.scaladsl.model.StatusCodes
-import com.google.common.primitives.Ints
-import com.spotify.docker.client.messages.Network
+import com.dimafeng.testcontainers.PostgreSQLContainer
 import com.typesafe.config.{Config, ConfigFactory}
-import com.wavesplatform.common.utils.EitherExt2
 import com.wavesplatform.dex.history.DBRecords.{EventRecord, OrderRecord}
 import com.wavesplatform.dex.history.HistoryRouter._
+import com.wavesplatform.dex.it.api.responses.dex.{OrderStatus, OrderStatusResponse}
+import com.wavesplatform.dex.it.docker.base.BaseContainer
+import com.wavesplatform.dex.model.MatcherModel.Normalization
 import com.wavesplatform.dex.model.OrderValidator
+import com.wavesplatform.dex.settings.PostgresConnection
 import com.wavesplatform.dex.settings.PostgresConnection._
-import com.wavesplatform.dex.settings.{OrderHistorySettings, PostgresConnection}
-import com.wavesplatform.it.api.SyncHttpApi._
-import com.wavesplatform.it.api.SyncMatcherHttpApi._
-import com.wavesplatform.it.sync.config.MatcherPriceAssetConfig._
-import com.wavesplatform.it.{DockerContainerLauncher, MatcherSuiteBase}
+import com.wavesplatform.it.MatcherSuiteBase
 import com.wavesplatform.transaction.Asset
+import com.wavesplatform.transaction.Asset.Waves
 import com.wavesplatform.transaction.assets.exchange.OrderType.{BUY, SELL}
 import com.wavesplatform.transaction.assets.exchange.{AssetPair, Order}
 import io.getquill.{PostgresJdbcContext, SnakeCase}
 import net.ceedubs.ficus.Ficus._
 
+import scala.concurrent.duration.DurationInt
 import scala.io.Source
-import scala.util.{Failure, Try}
+import scala.util.Try
 
 class PostgresHistoryDatabaseTestSuite extends MatcherSuiteBase {
 
-  // scenario:
-  //  1. create node network
-  //  2. create postgres container (pgc)
-  //  3. connect pgc to node network
-  //  4. up pgc
-  //  5. up node (during start node checks connection to postgres)
-  //  6. send messages from node to pgc
+  private val customDB       = "user_db"
+  private val customUser     = "user"
+  private val customPassword = "user"
 
-  val wavesNetwork: Network = dockerSingleton().createNetwork
+  private val postgresContainerName = "pgc"
+  private val postgresContainerPort = 5432
 
-  val postgresImageName, postgresUser = "postgres"
-  val postgresContainerName           = "pgc"
-  val postgresPassword                = "docker"
-  val postgresEnv                     = s"POSTGRES_PASSWORD=$postgresPassword"
-  val postgresContainerPort           = "5432"
-  val postgresContainerIp: String     = InetAddress.getByAddress(Ints.toByteArray(10 & 0xF | dockerSingleton().networkSeed)).getHostAddress
+  override protected val dexInitialSuiteConfig: Config = ConfigFactory.parseString(
+    s"""
+       |waves.dex {
+       |  price-assets = [ "$UsdId", "$BtcId", "WAVES", "$EthId", "$WctId" ]
+       |  ${getPostgresConnectionCfgString(postgresContainerName, postgresContainerPort)}
+       |  order-history.enabled = yes
+       |}
+    """.stripMargin
+  )
 
-  val postgresContainerLauncher =
-    new DockerContainerLauncher(
-      postgresImageName,
-      postgresContainerName,
-      List(postgresEnv),
-      postgresContainerIp,
-      postgresContainerPort,
-      wavesNetwork.name
-    )
+  private def getPostgresConnectionCfgString(serverName: String, port: Int): String =
+    s"""
+       |postgres {
+       |  server-name = $serverName
+       |  port-number = $port
+       |  database = $customDB
+       |  user = $customUser
+       |  password = $customPassword
+       |  data-source-class-name = "org.postgresql.ds.PGSimpleDataSource"
+       |}
+    """.stripMargin
 
-  val batchLingerMs: Int = OrderHistorySettings.defaultBatchLingerMs
-
-  def getPostgresContainerHostPort: String = postgresContainerLauncher.getHostPort.explicitGet()
-
-  @annotation.tailrec
-  private def retry[T](attemptsCount: Int, delayInMs: Int = 0)(fn: => T): Try[T] = {
-    Try { fn } match {
-      case Failure(_) if attemptsCount > 1 => if (delayInMs != 0) Thread.sleep(delayInMs); retry(attemptsCount - 1)(fn)
-      case Failure(ex)                     => throw ex
-      case success                         => success
+  private val postgres: PostgreSQLContainer =
+    PostgreSQLContainer(
+      databaseName = customDB,
+      username = customUser,
+      password = customPassword
+    ).configure { p =>
+      p.withNetwork(BaseContainer.network)
+      p.withNetworkAliases(postgresContainerName)
+      p.withCreateContainerCmdModifier { cmd =>
+        cmd withName postgresContainerName
+        cmd withIpv4Address BaseContainer.getIp(11)
+      }
     }
-  }
 
-  def createTables(postgresAddress: String): Unit = {
+  private def createTables(): Unit = {
 
-    val url                     = s"jdbc:postgresql://$postgresAddress/$postgresImageName"
     val orderHistoryDDLFileName = "/order-history/order-history-ddl.sql"
+
+    def getFileContentStr(fileName: String): String = {
+      val fileStream = getClass.getResourceAsStream(fileName)
+      Source.fromInputStream(fileStream).getLines.toSeq.mkString
+    }
 
     def executeCreateTablesStatement(sqlConnection: Connection): Try[Unit] = Try {
 
-      val fileStream            = getClass.getResourceAsStream(orderHistoryDDLFileName)
-      val createTablesDDL       = Source.fromInputStream(fileStream).getLines.toSeq.mkString
+      val createTablesDDL       = getFileContentStr(orderHistoryDDLFileName)
       val createTablesStatement = sqlConnection.prepareStatement(createTablesDDL)
 
       createTablesStatement.executeUpdate()
       createTablesStatement.close()
     }
 
-    retry(10, 2000) { DriverManager.getConnection(url, postgresUser, postgresPassword) } flatMap { sqlConnection =>
-      executeCreateTablesStatement(sqlConnection).map(_ => sqlConnection.close())
-    } get
+    val sqlConnection = DriverManager.getConnection(postgres.jdbcUrl, postgres.username, postgres.password)
+    executeCreateTablesStatement(sqlConnection).map(_ => sqlConnection.close())
   }
-
-  def getPostgresConnectionCfgString(serverName: String, port: String): String =
-    s"""
-       |postgres {
-       |  server-name = $serverName
-       |  port-number = $port
-       |  database = $postgresUser
-       |  user = $postgresUser
-       |  password = $postgresPassword
-       |  data-source-class-name = "org.postgresql.ds.PGSimpleDataSource"
-       |}
-    """.stripMargin
-
-  def getOrdersHistoryCfgString(batchLingerMs: Long): String =
-    s"""
-       |waves.dex {
-       |  ${getPostgresConnectionCfgString(postgresContainerName, postgresContainerPort)}
-       |  order-history {
-       |    enabled = yes
-       |    orders-batch-linger-ms = $batchLingerMs
-       |    orders-batch-entries = 10000
-       |    events-batch-linger-ms = $batchLingerMs
-       |    events-batch-entries = 10000
-       |  },
-       |  allowed-order-versions = [1, 2, 3]
-       |}
-    """.stripMargin
-
-  override protected def nodeConfigs: Seq[Config] =
-    super.nodeConfigs.map(
-      ConfigFactory
-        .parseString(getOrdersHistoryCfgString(batchLingerMs))
-        .withFallback
-    )
 
   override protected def beforeAll(): Unit = {
-    super.beforeAll()
+    // DEX depends on Postgres, so it must start before
+    postgres.start()
+    createTables()
 
-    postgresContainerLauncher.startContainer()
-    createTables(s"localhost:$getPostgresContainerHostPort")
+    wavesNode1.start()
 
-    Seq(IssueUsdTx, IssueWctTx, IssueEthTx, IssueBtcTx).foreach { tx =>
-      node.waitForTransaction { node.broadcastRequest(tx.json.value).id }
-    }
+    broadcastAndAwait(IssueUsdTx, IssueWctTx, IssueEthTx, IssueBtcTx)
 
-    node.upsertRate(eth, 0.00567593, expectedStatusCode = StatusCodes.Created)
-    node.upsertRate(btc, 0.00009855, expectedStatusCode = StatusCodes.Created)
+    dex1.start()
+
+    dex1.api.upsertRate(eth, 0.00567593)
+    dex1.api.upsertRate(btc, 0.00009855)
+    dex1.api.upsertRate(usd, 0.5)
   }
 
-  override protected def afterAll(): Unit = {
-    postgresContainerLauncher.stopAndRemoveContainer()
-    super.afterAll()
-  }
-
-  lazy val ctx =
+  private lazy val ctx =
     new PostgresJdbcContext(
       SnakeCase,
       ConfigFactory
-        .parseString(getPostgresConnectionCfgString("localhost", getPostgresContainerHostPort))
+        .parseString(getPostgresConnectionCfgString("localhost", postgres mappedPort postgresContainerPort))
         .as[PostgresConnection]("postgres")
         .getConfig
     )
 
   import ctx._
 
-  def getOrdersCount: Long = ctx.run(querySchema[OrderRecord]("orders", _.id      -> "id").size)
-  def getEventsCount: Long = ctx.run(querySchema[EventRecord]("events", _.orderId -> "order_id").size)
+  private def getOrdersCount: Long = ctx.run(querySchema[OrderRecord]("orders", _.id      -> "id").size)
+  private def getEventsCount: Long = ctx.run(querySchema[EventRecord]("events", _.orderId -> "order_id").size)
 
-  case class OrderBriefInfo(id: String,
-                            senderPublicKey: String,
-                            side: Byte,
-                            amountAsset: String,
-                            priceAsset: String,
-                            feeAsset: String,
-                            amount: Double,
-                            price: Double,
-                            fee: Double)
+  private case class OrderBriefInfo(id: String,
+                                    tpe: Byte,
+                                    senderPublicKey: String,
+                                    side: Byte,
+                                    amountAsset: String,
+                                    priceAsset: String,
+                                    feeAsset: String,
+                                    amount: Double,
+                                    price: Double,
+                                    fee: Double)
 
-  case class EventBriefInfo(orderId: String,
-                            eventType: Byte,
-                            filled: Double,
-                            totalFilled: Double,
-                            feeFilled: Double,
-                            feeTotalFilled: Double,
-                            status: Byte)
+  private case class EventBriefInfo(orderId: String,
+                                    eventType: Byte,
+                                    filled: Double,
+                                    totalFilled: Double,
+                                    feeFilled: Double,
+                                    feeTotalFilled: Double,
+                                    status: Byte)
 
-  def getOrderInfoById(orderId: String): Option[OrderBriefInfo] =
+  private def getOrderInfoById(orderId: Order.Id): Option[OrderBriefInfo] =
     ctx
       .run(
         querySchema[OrderBriefInfo](
           "orders",
           _.id              -> "id",
+          _.tpe             -> "type",
           _.senderPublicKey -> "sender_public_key",
           _.side            -> "side",
           _.amountAsset     -> "amount_asset_id",
@@ -187,24 +155,39 @@ class PostgresHistoryDatabaseTestSuite extends MatcherSuiteBase {
           _.amount          -> "amount",
           _.price           -> "price",
           _.fee             -> "fee"
-        ).filter(_.id == lift(orderId))
+        ).filter(_.id == lift(orderId.toString))
       )
       .headOption
 
-  def getEventsInfoByOrderId(orderId: String): Set[EventBriefInfo] =
+  private def getEventsInfoByOrderId(orderId: Order.Id): List[EventBriefInfo] =
     ctx
       .run(
-        querySchema[EventBriefInfo](
+        querySchema[EventRecord](
           "events",
           _.eventType      -> "event_type",
           _.filled         -> "filled",
           _.totalFilled    -> "total_filled",
           _.feeFilled      -> "fee_filled",
           _.feeTotalFilled -> "fee_total_filled",
-          _.status         -> "status"
-        ).filter(_.orderId == lift(orderId))
+          _.status         -> "status",
+          _.timestamp      -> "timestamp"
+        ).filter(_.orderId == lift(orderId.toString))
       )
-      .toSet
+      .sortWith { (l, r) =>
+        ((l.timestamp isBefore r.timestamp) || (l.timestamp isEqual r.timestamp)) &&
+        (l.status <= r.status) &&
+        (l.totalFilled <= r.totalFilled)
+      }
+      .map { r =>
+        EventBriefInfo(r.orderId, r.eventType, r.filled.toDouble, r.totalFilled.toDouble, r.feeFilled.toDouble, r.feeTotalFilled.toDouble, r.status)
+      }
+
+  implicit class DoubleOps(value: Double) {
+    val wct, usd: Long      = Normalization.normalizeAmountAndFee(value, 2)
+    val eth, btc: Long      = Normalization.normalizeAmountAndFee(value, 8)
+    val wctUsdPrice: Long   = Normalization.normalizePrice(value, 2, 2)
+    val wavesUsdPrice: Long = Normalization.normalizePrice(value, 8, 2)
+  }
 
   def stringify(asset: Asset): String = AssetPair.assetIdStr(asset)
 
@@ -212,12 +195,12 @@ class PostgresHistoryDatabaseTestSuite extends MatcherSuiteBase {
     val ordersCount = OrderValidator.MaxActiveOrders
 
     (1 to ordersCount)
-      .foreach { _ =>
-        node.placeOrder(alice, wctUsdPair, BUY, 1.wct, 0.35.price, 0.003.waves)
-        node.placeOrder(bob, wctUsdPair, SELL, 1.wct, 0.35.price, 0.003.waves)
+      .foreach { i =>
+        dex1.api.place(mkOrder(alice, wctUsdPair, BUY, 1.wct, 0.35.wctUsdPrice, 0.003.waves, ttl = 1.day + i.seconds))
+        dex1.api.place(mkOrder(bob, wctUsdPair, SELL, 1.wct, 0.35.wctUsdPrice, 0.003.waves, ttl = 1.day + i.seconds))
       }
 
-    retry(10, batchLingerMs) {
+    eventually {
       getOrdersCount shouldBe ordersCount * 2
       getEventsCount shouldBe ordersCount * 2
     }
@@ -225,76 +208,242 @@ class PostgresHistoryDatabaseTestSuite extends MatcherSuiteBase {
 
   "Order history should correctly save events: 1 big counter and 2 small submitted" in {
 
-    def sellOrder: Order = node.prepareOrder(bob, wctUsdPair, SELL, 100.wct, 0.35.price, fee = 0.00000030.btc, feeAssetId = btc, version = 3)
-    val buyOrder         = node.placeOrder(alice, wctUsdPair, BUY, 300.wct, 0.35.price, fee = 0.00001703.eth, feeAsset = eth, version = 3).message.id
+    def sellOrder: Order = mkOrder(bob, wctUsdPair, SELL, 100.wct, 0.35.wctUsdPrice, matcherFee = 0.00000030.btc, matcherFeeAssetId = btc)
+    val buyOrder         = mkOrder(alice, wctUsdPair, BUY, 300.wct, 0.35.wctUsdPrice, matcherFee = 0.00001703.eth, matcherFeeAssetId = eth)
 
-    val sellOrder1 = node.placeOrder(sellOrder).message.id
+    dex1.api.place(buyOrder)
 
-    node.waitOrderStatus(wctUsdPair, buyOrder, "PartiallyFilled")
-    node.waitOrderStatus(wctUsdPair, sellOrder1, "Filled")
+    val sellOrder1 = sellOrder
+    dex1.api.place(sellOrder1)
 
-    val sellOrder2 = node.placeOrder(sellOrder).message.id
+    dex1.api.waitForOrderStatus(buyOrder, OrderStatus.PartiallyFilled)
+    dex1.api.waitForOrderStatus(sellOrder1, OrderStatus.Filled)
 
-    node.waitOrderStatus(wctUsdPair, buyOrder, "PartiallyFilled")
-    node.waitOrderStatus(wctUsdPair, sellOrder2, "Filled")
+    val sellOrder2 = sellOrder
+    dex1.api.place(sellOrder2)
 
-    node.cancelOrder(alice, wctUsdPair, buyOrder)
+    dex1.api.waitForOrderStatus(buyOrder, OrderStatus.PartiallyFilled)
+    dex1.api.waitForOrderStatus(sellOrder2, OrderStatus.Filled)
 
-    retry(10, batchLingerMs) {
+    dex1.api.cancel(alice, buyOrder)
 
+    eventually {
       withClue("checking info for 2 small submitted orders\n") {
 
-        Set(sellOrder1, sellOrder2).foreach { orderId =>
-          getOrderInfoById(orderId).get shouldBe
-            OrderBriefInfo(orderId, bob.publicKey.toString, sellSide, stringify(wct), stringify(usd), stringify(btc), 100, 0.35, 0.00000030)
+        Set(sellOrder1, sellOrder2).foreach { order =>
+          getOrderInfoById(order.id()).get should matchTo(
+            OrderBriefInfo(order.idStr(),
+                           limitOrderType,
+                           bob.publicKey.toString,
+                           sellSide,
+                           stringify(wct),
+                           stringify(usd),
+                           stringify(btc),
+                           100,
+                           0.35,
+                           0.00000030)
+          )
 
-          getEventsInfoByOrderId(orderId) shouldBe Set { EventBriefInfo(orderId, eventTrade, 100, 100, 0.00000030, 0.00000030, statusFilled) }
+          getEventsInfoByOrderId(order.id()) should matchTo(
+            List(EventBriefInfo(order.idStr(), eventTrade, 100, 100, 0.00000030, 0.00000030, statusFilled))
+          )
         }
       }
 
       withClue("checking info for 1 big counter order\n") {
-        getOrderInfoById(buyOrder).get shouldBe
-          OrderBriefInfo(buyOrder, alice.publicKey.toString, buySide, stringify(wct), stringify(usd), stringify(eth), 300, 0.35, 0.00001703)
+        getOrderInfoById(buyOrder.id()).get should matchTo(
+          OrderBriefInfo(buyOrder.idStr(),
+                         limitOrderType,
+                         alice.publicKey.toString,
+                         buySide,
+                         stringify(wct),
+                         stringify(usd),
+                         stringify(eth),
+                         300,
+                         0.35,
+                         0.00001703)
+        )
 
-        getEventsInfoByOrderId(buyOrder) shouldBe
-          Set(
-            EventBriefInfo(buyOrder, eventTrade, 100, 100, 0.00000567, 0.00000567, statusPartiallyFilled),
-            EventBriefInfo(buyOrder, eventTrade, 100, 200, 0.00000567, 0.00001134, statusPartiallyFilled),
-            EventBriefInfo(buyOrder, eventCancel, 0, 200, 0, 0.00001134, statusCancelled)
+        getEventsInfoByOrderId(buyOrder.id()) should matchTo(
+          List(
+            EventBriefInfo(buyOrder.idStr(), eventTrade, 100, 100, 0.00000567, 0.00000567, statusPartiallyFilled),
+            EventBriefInfo(buyOrder.idStr(), eventTrade, 100, 200, 0.00000567, 0.00001134, statusPartiallyFilled),
+            EventBriefInfo(buyOrder.idStr(), eventCancel, 0, 200, 0, 0.00001134, statusCancelled)
           )
+        )
+      }
+    }
+  }
+
+  "Order history should correctly save events with Waves as amount and fee" in {
+    val buyOrder  = mkOrder(alice, wavesUsdPair, BUY, 300.waves, 0.35.wavesUsdPrice, matcherFee = 0.00370300.waves, matcherFeeAssetId = Waves)
+    val sellOrder = mkOrder(bob, wavesUsdPair, SELL, 300.waves, 0.35.wavesUsdPrice, matcherFee = 0.30.usd, matcherFeeAssetId = usd)
+
+    dex1.api.place(buyOrder)
+    dex1.api.place(sellOrder)
+
+    dex1.api.waitForOrderStatus(buyOrder, OrderStatus.Filled)
+    dex1.api.waitForOrderStatus(sellOrder, OrderStatus.Filled)
+
+    eventually {
+      withClue("checking info for counter order\n") {
+        getOrderInfoById(buyOrder.id()).get should matchTo(
+          OrderBriefInfo(buyOrder.idStr(), limitOrderType, alice.publicKey.toString, buySide, "WAVES", stringify(usd), "WAVES", 300, 0.35, 0.00370300)
+        )
+        getEventsInfoByOrderId(buyOrder.id()) should matchTo(
+          List(EventBriefInfo(buyOrder.idStr(), eventTrade, 300, 300, 0.00370300, 0.00370300, statusFilled))
+        )
+      }
+
+      withClue("checking info for submitted order\n") {
+        getOrderInfoById(sellOrder.id()).get should matchTo(
+          OrderBriefInfo(sellOrder.idStr(),
+                         limitOrderType,
+                         bob.publicKey.toString,
+                         sellSide,
+                         "WAVES",
+                         stringify(usd),
+                         stringify(usd),
+                         300,
+                         0.35,
+                         0.30)
+        )
+
+        getEventsInfoByOrderId(sellOrder.id()) should matchTo(
+          List(EventBriefInfo(sellOrder.idStr(), eventTrade, 300, 300, 0.30, 0.30, statusFilled))
+        )
       }
     }
   }
 
   "Order history should correctly save events: 1 small counter and 1 big submitted" in {
 
-    val smallBuyOrder = node.placeOrder(alice, wctUsdPair, BUY, 300.wct, 0.35.price, fee = 0.00001703.eth, feeAsset = eth, version = 3).message.id
-    val bigSellOrder  = node.placeOrder(bob, wctUsdPair, SELL, 900.wct, 0.35.price, fee = 0.00000030.btc, feeAsset = btc, version = 3).message.id
+    val smallBuyOrder = mkOrder(alice, wctUsdPair, BUY, 300.wct, 0.35.wctUsdPrice, 0.00001703.eth, matcherFeeAssetId = eth)
+    val bigSellOrder  = mkOrder(bob, wctUsdPair, SELL, 900.wct, 0.35.wctUsdPrice, 0.00000030.btc, matcherFeeAssetId = btc)
 
-    node.waitOrderStatus(wctUsdPair, smallBuyOrder, "Filled")
-    node.waitOrderStatus(wctUsdPair, bigSellOrder, "PartiallyFilled")
+    dex1.api.place(smallBuyOrder)
+    dex1.api.place(bigSellOrder)
 
-    retry(20, batchLingerMs) {
+    dex1.api.waitForOrderStatus(smallBuyOrder, OrderStatus.Filled)
+    dex1.api.waitForOrderStatus(bigSellOrder, OrderStatus.PartiallyFilled)
 
+    eventually {
       withClue("checking info for small counter order\n") {
-        getOrderInfoById(smallBuyOrder).get shouldBe
-          OrderBriefInfo(smallBuyOrder, alice.publicKey.toString, buySide, stringify(wct), stringify(usd), stringify(eth), 300, 0.35, 0.00001703)
+        getOrderInfoById(smallBuyOrder.id()).get should matchTo(
+          OrderBriefInfo(smallBuyOrder.idStr(),
+                         limitOrderType,
+                         alice.publicKey.toString,
+                         buySide,
+                         stringify(wct),
+                         stringify(usd),
+                         stringify(eth),
+                         300,
+                         0.35,
+                         0.00001703)
+        )
 
-        getEventsInfoByOrderId(smallBuyOrder) shouldBe Set {
-          EventBriefInfo(smallBuyOrder, eventTrade, 300, 300, 0.00001703, 0.00001703, statusFilled)
-        }
+        getEventsInfoByOrderId(smallBuyOrder.id()) should matchTo(
+          List(EventBriefInfo(smallBuyOrder.idStr(), eventTrade, 300, 300, 0.00001703, 0.00001703, statusFilled))
+        )
       }
 
       withClue("checking info for big submitted order\n") {
-        getOrderInfoById(bigSellOrder).get shouldBe
-          OrderBriefInfo(bigSellOrder, bob.publicKey.toString, sellSide, stringify(wct), stringify(usd), stringify(btc), 900, 0.35, 0.00000030)
+        getOrderInfoById(bigSellOrder.id()).get should matchTo(
+          OrderBriefInfo(bigSellOrder.idStr(),
+                         limitOrderType,
+                         bob.publicKey.toString,
+                         sellSide,
+                         stringify(wct),
+                         stringify(usd),
+                         stringify(btc),
+                         900,
+                         0.35,
+                         0.00000030)
+        )
 
-        getEventsInfoByOrderId(bigSellOrder) shouldBe Set {
-          EventBriefInfo(bigSellOrder, eventTrade, 300, 300, 0.00000010, 0.00000010, statusPartiallyFilled)
-        }
+        getEventsInfoByOrderId(bigSellOrder.id()) should matchTo(
+          List(EventBriefInfo(bigSellOrder.idStr(), eventTrade, 300, 300, 0.00000010, 0.00000010, statusPartiallyFilled))
+        )
       }
     }
-    node.cancelAllOrders(alice)
-    node.cancelAllOrders(bob)
+  }
+
+  "Order history should correctly save market orders and their events" in {
+
+    dex1.api.cancelAll(bob)
+    dex1.api.cancelAll(alice)
+
+    def bigBuyOrder: Order = mkOrder(alice, wctUsdPair, BUY, 500.wct, 0.35.wctUsdPrice, matcherFee = 0.00001703.eth, matcherFeeAssetId = eth)
+
+    withClue("place buy market order into empty order book") {
+
+      val unmatchableMarketBuyOrder = bigBuyOrder
+      dex1.api.placeMarket(unmatchableMarketBuyOrder)
+      dex1.api.waitForOrder(unmatchableMarketBuyOrder)(
+        _ == OrderStatusResponse(OrderStatus.Filled, filledAmount = Some(0.wct), filledFee = Some(0.wct)))
+
+      eventually {
+        getOrderInfoById(unmatchableMarketBuyOrder.id()).get should matchTo(
+          OrderBriefInfo(unmatchableMarketBuyOrder.idStr(),
+                         marketOrderType,
+                         alice.publicKey.toString,
+                         buySide,
+                         stringify(wct),
+                         stringify(usd),
+                         stringify(eth),
+                         500,
+                         0.35,
+                         0.00001703)
+        )
+
+        getEventsInfoByOrderId(unmatchableMarketBuyOrder.id()) should matchTo(
+          List(EventBriefInfo(unmatchableMarketBuyOrder.idStr(), eventCancel, 0, 0, 0, 0, statusFilled))
+        )
+      }
+    }
+
+    withClue("place buy market order into nonempty order book") {
+
+      val ts = System.currentTimeMillis()
+
+      val orders = Seq(
+        mkOrder(bob, wctUsdPair, SELL, 100.wct, 0.33.wctUsdPrice, 0.003.waves, ts = ts),
+        mkOrder(bob, wctUsdPair, SELL, 100.wct, 0.34.wctUsdPrice, 0.003.waves, ts = ts + 100),
+        mkOrder(bob, wctUsdPair, SELL, 100.wct, 0.34.wctUsdPrice, 0.003.waves, ts = ts + 200)
+      )
+
+      orders.foreach { order =>
+        dex1.api.place(order)
+        dex1.api.waitForOrderStatus(order, OrderStatus.Accepted)
+      }
+
+      val marketBuyOrder = bigBuyOrder
+      dex1.api.placeMarket(marketBuyOrder)
+      dex1.api.waitForOrder(marketBuyOrder)(_ == OrderStatusResponse(OrderStatus.Filled, filledAmount = Some(300.wct), filledFee = Some(0.003.waves)))
+
+      eventually {
+        getOrderInfoById(marketBuyOrder.id()).get should matchTo(
+          OrderBriefInfo(marketBuyOrder.idStr(),
+                         marketOrderType,
+                         alice.publicKey.toString,
+                         buySide,
+                         stringify(wct),
+                         stringify(usd),
+                         stringify(eth),
+                         500,
+                         0.35,
+                         0.00001703)
+        )
+
+        getEventsInfoByOrderId(marketBuyOrder.id()) should matchTo(
+          List(
+            EventBriefInfo(marketBuyOrder.idStr(), eventTrade, 100, 100, 0.00000340, 0.00000340, statusPartiallyFilled),
+            EventBriefInfo(marketBuyOrder.idStr(), eventTrade, 100, 200, 0.00000340, 0.00000680, statusPartiallyFilled),
+            EventBriefInfo(marketBuyOrder.idStr(), eventTrade, 100, 300, 0.00000340, 0.00001020, statusPartiallyFilled),
+            EventBriefInfo(marketBuyOrder.idStr(), eventCancel, 0, 300, 0, 0.00001020, statusFilled)
+          )
+        )
+      }
+    }
   }
 }
