+++ dex-it/src/test/scala/com/wavesplatform/it/sync/orders/OrderFixedFeeTestSuite.scala
@@ -2,13 +2,10 @@ package com.wavesplatform.it.sync.orders
 
 import com.typesafe.config.Config
 import com.typesafe.config.ConfigFactory.parseString
-import com.wavesplatform.account.{KeyPair, PublicKey}
+import com.wavesplatform.account.PublicKey
 import com.wavesplatform.common.state.ByteStr
 import com.wavesplatform.common.utils.EitherExt2
-import com.wavesplatform.it.api.SyncHttpApi._
-import com.wavesplatform.it.api.SyncMatcherHttpApi._
-import com.wavesplatform.it.sync.config.MatcherPriceAssetConfig.{matcher, _}
-import com.wavesplatform.it.{MatcherSuiteBase, NTPTime}
+import com.wavesplatform.it.MatcherSuiteBase
 import com.wavesplatform.lang.script.Script
 import com.wavesplatform.lang.script.v1.ExprScript
 import com.wavesplatform.lang.v1.compiler.Terms
